+++ fe/src/components/Header/index.tsx
@@ -2,13 +2,13 @@ import { useState } from 'react';
 import { useLocation } from 'react-router-dom';
 
 import styled, { css } from 'styled-components';
-import { Avatar, Link } from '@mui/material';
-import { ReactComponent as HamburgerIcon } from 'images/FE_숙소예약서비스/Property 1=menu.svg';
+import { Link } from '@mui/material';
 import { ReactComponent as LogoImg } from 'images/logo.svg';
 
 import Menu from 'components/Header/Menu';
 import SearchBar from 'components/Header/SearchBar/index';
 import SearchWrapper, { SearchWrap } from './SearchWrapper';
+import UserMenu from './UserMenu';
 
 // type Router = '/' | '/result';
 
@@ -41,10 +41,7 @@ function Header() {
           setMiniFocus={setMiniFocus}
         />
       )}
-      <UserWrapper type="button" aria-label="유저 정보 메뉴">
-        <Hamburger />
-        <CustomAvatar src="images/FE_숙소예약서비스/Property 1=user.svg" />
-      </UserWrapper>
+      <UserMenu />
     </HeaderWrap>
   );
 }
@@ -72,32 +69,4 @@ const HeaderWrap = styled.header<Position>`
   z-index: 1;
 `;
 
-const UserWrapper = styled.button`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 76px;
-  height: 40px;
-  border-radius: 999px;
-  border: 1px solid ${({ theme }) => theme.colors.grey4};
-  background: ${({ theme }) => theme.colors.white};
-`;
-
-const Hamburger = styled(HamburgerIcon)`
-  width: 16px;
-  height: 16px;
-  margin: 0 8px 0 10px;
-
-  path {
-    stroke: ${({ theme }) => theme.colors.black};
-  }
-`;
-
-const CustomAvatar = styled(Avatar)`
-  && {
-    width: 32px;
-    height: 32px;
-    background: ${({ theme }) => theme.colors.grey3};
-  }
-`;
 export default Header;
